Server/requirements.txt
absl-py==0.11.0
asgiref==3.3.1
astunparse==1.6.3
cachetools==4.2.0
certifi==2020.12.5
chardet==3.0.4
cycler==0.10.0
decorator==4.4.2
Django==3.0.5
django-rest-framework==0.1.0
djangorestframework==3.12.2
djongo==1.3.3
filelock==3.0.12
gast==0.3.3
gdown==3.11.1
gluoncv==0.7.0
google-auth==1.24.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
graphviz==0.8.4
grpcio==1.34.0
h5py==2.10.0
idna==2.6
imageio==2.9.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.3
matplotlib==3.3.3
mxnet==1.6.0
networkx==2.5
numpy==1.16.6
oauthlib==3.1.0
opencv-contrib-python==4.2.0.34
opt-einsum==3.3.0
Pillow==7.1.2
portalocker==2.0.0
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pymongo==3.11.2
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2020.5
PyWavelets==1.1.1
pywebview==3.2
requests-oauthlib==1.3.0
rsa==4.6
scikit-image==0.15.0
scipy==1.4.1
six==1.15.0
sqlparse==0.2.4
tensorboard==2.2.2
tensorboard-plugin-wit==1.7.0
tensorflow==2.2.1
tensorflow-estimator==2.2.0
termcolor==1.1.0
torch==1.7.1
torchvision==0.2.2.post3
tqdm==4.43.0
typing-extensions==3.7.4.3
urllib3==1.22
Werkzeug==1.0.1
wrapt==1.12.1
